PT-3934: Make internal diagnosis prediction tool default, rather than BOQA
